models can just live alongside services
