--- backend/src/models/qrCode.ts
@@ -1,4 +1,0 @@
-export type QrCode = {
-    qid: number
-    dishId: string
-}
