--- backend/src/models/user.ts
@@ -1,5 +1,0 @@
-export type User = {
-    id: string
-    role: string
-    email: string
-}
